Add files via upload
Survey form to gather information from the users. This could be feedback, information from
customers, or data for research purposes. You employ HTML to build the structure of your survey
form and to create the fields the users have to fill (name, age, gender, details of contact). CSS can
do the styling. JavaScript is for tasks associated with validation such as e-mail ID and character
limit.  https://github.com/AscentAcademy